+++ Memoria.Scripts/Sources/Battle/0105_GrandCrossScript.cs
@@ -41,7 +41,7 @@ namespace Memoria.Scripts.Battle
                 else if ((status & (UInt32)BattleStatus.LowHP) != 0 && !_v.Target.IsUnderStatus(BattleStatus.Death))
                 {
                     _v.Context.Flags |= BattleCalcFlags.DirectHP;
-                    _v.Target.CurrentHp = (UInt16)(1 + GameRandom.Next8() % 9);
+                    _v.Target.CurrentHp = (UInt32)(1 + GameRandom.Next8() % 9);
                 }
             }
         }
